Make frame count ulong.

+++ ts-split.c
@@ -43,7 +43,7 @@ typedef struct {
 typedef struct {
   AVFormatContext *file;
   input_stream **st;
-  int frame_count;
+  unsigned long frame_count;
   FILE *mf;
 } tss_input;
 
@@ -52,7 +52,7 @@ typedef struct {
   output_stream **st;
   char *tmp_name;
   char *name;
-  int first_frame;
+  unsigned long first_frame;
 } tss_output;
 
 static int verbose = 0;
@@ -407,8 +407,8 @@ static void
 post_command( const char *cmd, tss_output * out, tss_input * in ) {
   if ( cmd ) {
     char sbuf[32], ebuf[32];
-    sprintf( sbuf, "%d", out->first_frame );
-    sprintf( ebuf, "%d", in->frame_count );
+    sprintf( sbuf, "%lu", out->first_frame );
+    sprintf( ebuf, "%lu", in->frame_count );
     const char *dict[] = { CHUNK_NAME, CHUNK_START, CHUNK_END, NULL };
     const char *vals[] = { out->name, sbuf, ebuf, NULL };
     char *cmdbuf = str_replace_multi( cmd, dict, vals );
@@ -488,7 +488,7 @@ start_output( tss_output * out, tss_input * in, const char *name, int seq ) {
 
 static void
 write_manifest( FILE * fl, tss_output * out, tss_input * in ) {
-  fprintf( fl, "%s,%d,%d\n", out->name, out->first_frame,
+  fprintf( fl, "%s,%lu,%lu\n", out->name, out->first_frame,
            in->frame_count );
 }
